rqs/requirements.txt
asgiref==3.2.10
autopep8==1.5.4
beautifulsoup4==4.8.2
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
dj-database-url==0.5.0
Django==3.0.9
django-appconf==1.0.4
django-bootstrap-modal-forms==2.0.0
django-colorful==1.3
django-el-pagination==3.3.0
django-filter==2.3.0
django-js-asset==1.2.2
django-modelcluster==5.0.2
django-phone-field==1.8.1
django-phonenumber-field==5.0.0
django-social-share==2.0.0
django-taggit==1.3.0
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.1
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
flake8==3.8.3
gunicorn==20.0.4
html5lib==1.1
idna==2.10
jdcal==1.4.1
l18n==2018.5
mccabe==0.6.1
numpy==1.19.1
openpyxl==3.0.5
pandas==1.1.1
phonenumbers==8.12.8
Pillow==7.2.0
plotly==4.9.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
retrying==1.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.1
Unidecode==1.1.1
Unipath==1.1
urllib3==1.25.10
wagtail==2.10.1
webencodings==0.5.1
whitenoise==5.2.0
Willow==1.4
xlrd==1.2.0
XlsxWriter==1.3.3
xlwt==1.3.0
